Keep atomic writer dependencies out of ordinary file-manager clients
