+++ test/SConstruct
@@ -44,7 +44,7 @@ if sys.platform == "darwin" or sys.platform.startswith("linux"):
 gag_server = gag_env.StaticLibrary('gagserver', [
   gag_env.Object('gagserver-' + name + '.o', '../libgag/src/' + name + '.cpp')
   for name in Split("""
-    BinaryStream FileManager FormatableString Stream StreamBackend
+    BinaryStream FileManager FileManagerAtomic FormatableString Stream StreamBackend
     StreamFilter StringTable TextStream Toolkit
   """)])
